Version 1.8
  * More cross-platform services: keepalived, dovecot, wicd
  * More BSD services: ccd, timed, mroute4d, mroute6d, rwhod, ipxrouted, nfscbd, nfsuserd, bootparams
  * system-control now has a built-in version subcommand.
  * Fixed a bug with convert-systemd-units and Environment= settings that attempt to unset environment variables.
  * Added an ExecRestartPre= setting to convert-systemd-units.
  * Fixed a bug in the regular-services postrm script.
  * Fixed a bug that prevented the fastXYZ system V shims from being recognized.
  * Documented the nagios-check-service command.

+++ source/pty-run.cpp
@@ -256,5 +256,5 @@ pty_run (
 		tcsetattr_nointr(STDIN_FILENO, TCSADRAIN, original_attr);
 	if (!WIFEXITED(status))
 		waitpid(child, &status, 0);
-	throw WIFEXITED(status) ? WEXITSTATUS(status) : EXIT_TEMPORARY_FAILURE;
+	throw WIFEXITED(status) ? WEXITSTATUS(status) : static_cast<int>(EXIT_TEMPORARY_FAILURE);
 }
